remove tune manager
