+++ scvi/autotune/__init__.py
@@ -3,7 +3,7 @@ from scvi.utils import error_on_missing_dependencies
 error_on_missing_dependencies("hyperopt", "ray.tune")
 
 
-from ._manager import TuneAnalysis, TunerManager  # noqa
+from ._experiment import Experiment  # noqa
 from ._tuner import ModelTuner  # noqa
 
-__all__ = ["ModelTuner", "TunerManager", "TuneAnalysis"]
+__all__ = ["Experiment", "ModelTuner"]
